separated subscriber-attributes-related tests for device cache into its own file
